requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
Pillow
imageio
httpx
# torch
# accelerate
# diffusers
# transformers